'Flatten' graph spikes
Fixed the issue with bandwidth graphs showing spikes at startup, which,
because of the high value of the spike, made the rest of the values
invisible for some time. Now the average value of the adjacent values
are replacing the spike value, if the spike value is greater than
the bandwidth burst value.